x86/watchdog/hardlockup/hpet: Determine if HPET timer caused NMI
The only direct method to determine whether an HPET timer caused an
interrupt is to read the Interrupt Status register. Unfortunately,
reading HPET registers is slow and, therefore, it is not recommended to
read them while in NMI context.

An indirect manner is to compute the expected value of the the time-stamp
counter and, at the time of the interrupt and verify that its actual
value is within a range of the expected value. Since the hardlockup
detector operates in seconds, high precision is not needed. This
implementation considers that the HPET caused the HMI if the time-stamp
counter reads the expected value -/+ 2%.

Also, to avoid variability across CPUs, always read the time-stamp counter
of the same CPU, the handling CPU. The handling CPU can only change when
the lockup detector is started or reconfigured.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>