x86/watchdog/hardlockup/hpet: Add support for interrupt remapping
When interrupt remapping is enabled in the system. The MSI interrupt
message must follow a special format the IOMMU can understand. Hence,
utilize the functionality provided by the IOMMU driver for such purpose.

The first step is to determine whether interrupt remapping is enabled
by looking for the existence of an interrupt remapping domain. If it
exists, let the IOMMU driver compose the MSI message for us. The hard
lockup detector is still responsible of writing the message in the
HPET FSB route register.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>