iommu/vt-d: hpet: Reserve an interrupt remampping table entry for watchdog
When interrupt remapping is enabled, MSI interrupt messages must follow a
special format that the IOMMU can understand. Hence, when the HPET hard
lockup detector is used with interrupt remapping, it must also follow this
specia format.

The IOMMU, given the information about a particular interrupt, populates
an entry in the interrupt remapping table. Add two functions to create
an entry for the HPET hard lockup detector. Perform this process in two
steps as described below.

When initializing the lockup detector, the function
watchdog_hld_hpet_alloc_irq() allocates a new entry in the interrupt
remapping table by populating data structures with the information the
IOMMU driver needs. In order to populate the table, the IOMMU needs to
know the HPET block ID as described in the ACPI table. Hence, add such
ID to the data of the hardlockup detector.

When the hardlockup detector is enabled, the function
watchdog_hld_hpet_activate_irq() activates the recently created entry
in the interrupt remapping table via the modify_irte() functions. While
doing this, specify which CPU the interrupt must target via its APIC ID.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>