x86/hpet: Introduce function to identify HPET hardlockup detector irq
The interrupt remapping driver relies on the irq subsystem to allocate
and activate interrupts. However, the delivery mode of the interrupts
is inherited from the APIC delivery mode.

The HPET hardlockup detector needs to deliver its interrupt as NMI.
Unfortunately, in x86 we don't have a IRQF_NMI flag that can be used
when requesting an interrupt (via request_irq()). Therefore, there is
not a way in which the irq subsystem can specify NMI delivery mode.

Instead, when allocating the irq, inspect the irq private data to
to determine whether it belongs to the HPET hardlockup detector. If this
is the case, update the delivery mode.

Cc: Ashok Raj <ashok.raj@intel.com>
Cc: Andi Kleen <andi.kleen@intel.com>
Cc: Tony Luck <tony.luck@intel.com>
Cc: Borislav Petkov <bp@suse.de>
Cc: Jacob Pan <jacob.jun.pan@intel.com>
Cc: Joerg Roedel <joro@8bytes.org>
Cc: Juergen Gross <jgross@suse.com>
Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: Wincy Van <fanwenyi0529@gmail.com>
Cc: Kate Stewart <kstewart@linuxfoundation.org>
Cc: Philippe Ombredanne <pombredanne@nexb.com>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Baoquan He <bhe@redhat.com>
Cc: Jan Kiszka <jan.kiszka@siemens.com>
Cc: Lu Baolu <baolu.lu@linux.intel.com>
Cc: Stephane Eranian <eranian@google.com>
Cc: Suravee Suthikulpanit <Suravee.Suthikulpanit@amd.com>
Cc: "Ravi V. Shankar" <ravi.v.shankar@intel.com>
Cc: x86@kernel.org
Cc: iommu@lists.linux-foundation.org
Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>